+++ main-service/src/main/java/ru/practicum/mainservice/request/model/IConfirmedRequests.java
@@ -2,5 +2,5 @@ package ru.practicum.mainservice.request.model;
 
 public interface IConfirmedRequests {
     Long getEventId();
-    Long getConfirmedCount();
+    int getConfirmedCount();
 }
